requirements.txt
setuptools
azure-search-documents>=11.4.0
openai
tenacity
tiktoken
bs4
packaging